Implement wave selection and single-bit toggle onClick
This change allows the user to click on a wave and select it, turning it
green. The card then implements a deselect function when the mouse
clicks outside of it. This can be used for other functionality in the
future but is now being used to deselect whichever wave is selected.
Once a wave is selected (indicated by a green background) the user can
click on a single-bit wave and the value under the mouse will be
toggled.