Update wave sequence check to happen immediately
Keep the delay to inform the user if the sequence is bad but immediately
check it. This way, as soon as the user types a correct sequence the red
error indication will go away. This makes it a little nicer for the
user. However, this change also fixes a bug where the sequence would be
empty if the user input an incorrect sequence, kept typing, and very
quickly hit add.