Fix bug with period input field
When editing a wave, if the user entered a new number in the period
field and hit enter to submit the form, the browser would complain about
an invalid number and say that the nearest valid values were +/- 0.5 of
the entered value. Adding the step='any' attribute to the period input
element fixed the issue.
This honestly still seems weird though because the user was still able
to hit the "add" button to submit the form and the value entered into
the period input was used correctly. Anyway, user no longer gets the
error message.