Fix crash when deleting a wave
The selectedWaveId is used to determine which wave to pass to the
WaveEditDrawer. When a wave was selected and then deleted, the
selectedWaveId did not change but we could no longer find a wave with
the appropriate ID to pass to the WaveEditDrawer. So things got nasty.
Fix was to set the SelectedWaveId to null upon delete. That should have
been there in the first place anyway.